can you move "Gate Pipe" over to be right next to Chain-Link, Also change the title of "Chain-Link" to just Chain Link
Also add in the option for Gate in every other category besides chain link. make it so you can pick either "Single Gates" which would give you the option of picking 3' , 4' , and 5' and then an "Double Gates" With the quantity's of the single gate doubled. if Gates are picked for vinyl, Single gates should add one more end post and a gate post, as well as the hardware needed for it, SS PVC hinge (1) SS PVC Latch (1) or for a double - SS PVC Hinge (2), SS PVC Latch (1) SS Drop Rod.