app is working great and everything looks pretty good, but i do have a few tweaks.
under all of the category's, can we add an option to have more than just one gate?

for the vinyl section, If the footage is more than say 2 1/2 sections, Round up instead of rounding down.

Under unit converter, can we add a tool labled "Lakeland 2" and give me the option to input the amount of sections we need, in turn it will spit out a breakdown of the materials needed (1 section = 13 Pickets, 2 Rails, 2 U-Channels)